Angularfy your web page
Within our app.js file we need to instantiate an AngularJS module which
we will use for our 'app' declared in index.html
<body ng-app='app'>

The angular.module() function acts as a setter and getter for angular
modules. The first argument is the name of the module where we chose
app, for the second we provided an array. If an array is provided then
angular will create the app module as opposed to retrieving it which
you’ll see later when we create our controller.

In more complex applications you will provide an array of module
dependencies as opposed to an empty array as with this example.