The Controller for app.js
AngularJS provides a loose MVC structure more accurately described as
MVVM. The controller is where your business logic should be handled,
essentially the controller ‘controls’ the data and provides it to the
view for display.

The module 'app' is declared in app.js.  app.js is called from
index.html

In the code above we first retrieve the app module we created early and
then use the controller function to instantiate a new controller. The
controller function takes two parameters. The first is the controller
name and the second is a function where we will place our controlling
code.

Before we can use the controller we need to tell angular which part of
our HTML document this controller has governance over. You can have many
controllers per document and many controllers governing the same section
of a document, you can even have nested controllers.

In our very simple example we are going to have a single controller
governing the entire page. We will use the controllerAs syntax to
declare this (more on this later), open index.html and add the following
to the body tag.