Here to show one controller to control one page
Before we can use the controller we need to tell angular which part of
our HTML document this controller has governance over. You can have many
controllers per document and many controllers governing the same section
of a document, you can even have nested controllers. In our very simple
example we are going to have a single controller governing the entire
page. We will use the controllerAs syntax to declare this (more on this
later), open index.html and add the following to the body tag.