To display the text 'AngularJS Tutorial Example' at main.ctrl.js to index.html
<div class="container">
<h1>{{main.title}}</h1>
</div>

In the above example we are displaying the title text inside a h1 tag
using a set of curly braces so that angular knows that this is a
variable. We access the controllers scope by prefixing the variable with
main matching our ControllerAs definition. Now imagine that we had
multiple controllers in the same document and you were referencing
multiple variables using the traditional method where you would simple
do {{title}}. It would be very difficult to see which scope the
variables were present in. Additionally it would also be easy to
accidentally have multiple variables with the same name (i.e scope
bleed).